refactored to use builder for git menu
